Crate qor_rto

Runtime Orchestration (RTO) is the central runtime control element of the Qore stack

Overview

Examples

use qor_rto::prelude::*;
use std::time::Duration;

// The Hello World Routine with the correct signature
fn hello_world(_delta: &Duration) -> (Duration, UpdateResult) {
    println!("Hello World!"); // Hello the world
    (Duration::ZERO, UpdateResult::Complete) // And declare the action as complete
}

// A Hello World program
fn main() {
    // The engine is the central runtime executor
    let engine = Engine::default();
    engine.start().unwrap();

    // Create a new program with a single invoke action
    let program = Program::new().with_action(Invoke::new(hello_world));

    // Spawn the program on the engine
    let handle = program.spawn(&engine).unwrap();

    // Wait for the program to finish
    let _ = handle.join().unwrap();

    // Engine shutdown
    engine.shutdown().unwrap();
}

Further examples can be found in the examples directory of the repository.

Modules

builder

prelude

rto_errors

Structs

AccurateClock

The AccurateClock is a clock that uses a monotonous system counter with high resolution to determine the current timepoint.

Activity

An Activity is an action that controls the user-implemented Lifecycle model.

Add

Add is an RValue that combines two RValues with an addition. std::ops::Add must be defined for T.

And

And is an RValue that combines two RValues with a logical AND for bool and an arithmetical AND for scalars. std::ops::And must be defined for T.

Assign

Assign copies the value from an RValue evaluation into an LValue It is coherent with the = operator in Rust

Await

Await is an asynchronous call into a user-defined function that acts as a coroutine.

Break

Break action exits the loop it is in.

Computation

A Computation is an action that executes multiple actions concurrently. It completes

Concurrency

A Concurrency is an action that executes other actions in multiple branches concurrently. It completes when all concurrent actions have completed. One branch throwing an exception will cause the Concurrency to throw the exception as well.

Constant

Constant is an RValue that always evaluates to the same value

Div

Div is an RValue that combines two RValues with an division. std::ops::Div must be defined for T.

Engine

An engine runs tasks concurrently on threads. It is the central executor of runtime orchestration.

EngineBuilder

The EngineBuilder configures an Engine as Executor

Eq

Eq is an RValue that compares two RValues and evaluates to true if both are equal std::cmp::PartialEq must be defined for T.

Event

An Event signals the occurrence of the change of state.

ExecutionContext

The context of a program’s execution.

ExecutionException

An ExecutionException represents a single expection thrown during execution There are 64 exception codes available for exceptions, thereof 56 are user defined exceptions.

ExecutionExceptionFilter

A filter for multiple exceptions

Fiber

A fiber is a lightweight cooperative thread running in a program.

For

For is a finite loop that loops it’s body for a definite number of times A Break or an exception can terminate the loop early. A For loop without a body will complete immediately. A For loop that contains neither Sleep, Sync or Await will never become busy.

ForRange

ForRange is a finite loop that loops a fixed number of times without an explicit iterator, start and end value.

Gt

Gt is an RValue that compares two RValues and evaluates to true if lhs is greater than rhs std::cmp::PartialOrd must be defined for T.

Gte

Gte is an RValue that compares two RValues and evaluates to true if lhs is greater or equal than rhs std::cmp::PartialOrd must be defined for T.

IfThenElse

IfThenElse is an action that executes a sequence of actions in order

Invoke

Update is a synchronous call into a user-defined function that acts as a coroutine.

IpcEvent

An event for a single listener

IpcListen

The listener for a single event

IpcNotify

Notify implementation for Ipc event

Listener

A Listener is a handle to an event to receive notifications

Loop

Loop is an infinite loop that loops it’s body until a Break or an exception terminates the loop.

Lt

Lt is an RValue that compares two RValues and evaluates to true if lhs is less than rhs std::cmp::PartialOrd must be defined for T.

Lte

Lte is an RValue that compares two RValues and evaluates to true if lhs is less or equal than rhs std::cmp::PartialOrd must be defined for T.

Mul

Mul is an RValue that combines two RValues with a multiplication. std::ops::Mul must be defined for T.

Neg

Neg is an RValue that negates another RValue std::ops::Neg must be defined for T.

Neq

Neq is an RValue that compares two RValues and evaluates to false if both are equal std::cmp::PartialEq must be defined for T.

Nop

Nop is a no-operation action that does nothing

Not

Not is an RValue that inverts another RValue std::ops::Not must be defined for T.

Notifier

A Notifier is a handle to an event to trigger notifications

Or

Or is an RValue that combines two RValues with a logical OR for bool and an arithmetical OR for scalars std::ops::Or must be defined for T.

Predicate

A Predicate is a user-defined function that evaluates to true or false

Program

A program is the central control element for running actions in fibers as tasks on an executor.

Sequence

Sequence is an action that executes a sequence of actions called step in consecutive order

SingleEvent

An event for a single listener

SingleListen

The listener for a single event

SingleNotify

Notify implementation for Single event

Sleep

Sleep is an action that waits for a specified duration

StateStorage

A StateStorage is a state with a variable value that implements the State trait and updates new values upon update only.

Sub

Sub is an RValue that combines two RValues with an substraction. std::ops::Sub must be defined for T.

Sync

Sync is an action that waits for a given Event to occur.

Task

A Task is the smallest unit of execution that can be scheduled on an executor.

TaskHandle

A TaskHandle is used to track a task after it has been spawned on an executor.

Throw

Throw action throws an exception with an optional timeout The timeout comes in handy when the throw is used in a parallel action in a Computation, thus controlling the main execution flow.

Trigger

Trigger is an action that triggers the given event

TryCatch

TryCatch is an action that executes a try action and catches exceptions with a given exception mask. The catch action is executed if an exception is thrown that matches the mask.

VariableStorage

A VariableStorage contains a single value storage fulfilling the Variable trait.

Xor

Xor is an RValue that combines two RValues with a logical XOR for bool and an arithmetical XOR for scalars /// std::ops::BitXor must be defined for T.

Enums

EvalError

Result of the evaluation of an RValue

ExecutionInterruption

Predefined exceptions for actions

ExecutionState

The ExecutionState enum represents the runtime state of executing elements. Executing elements are Actions, Programs, Tasks and Executors.

ExecutionTransition

The ExecutionTransition enum represents the possible transitions between execution states.

TimeoutStatus

UpdateResult

The ActionResult enum represents the result of the update function of an action.

Traits

Action

The Action trait describes the interface to an executable element we call action.

ActivityLifecycle

The lifecycle of an activity.

Clock

A clock provides the current timepoint and can hold the execution sleep for a given duration.

EventAdapter

An EventAdapter links an event implementation with the orchestration Event abstraction.

LValue

An LValue is a data element that can be assigned a value. The type T must comply to the Qore Type System.

Listen

Listen trait for event listeners.

Notify

The Notify trait is implemented by Notifiers

RValue

An RValue delivers a value of type T upon request. The type T must comply to the Qore Type System.

State

A State is a trait that represents a state storage with explicit timepoints for updates The type T must comply to the Qore Type System.

Variable

A Variable is a trait that represents a scalar variable storage with instant updates The type T must comply to the Qore Type System.

Type Aliases

ConditionType

ExecutionExceptionCode

Code type for execution exceptions Code ranges from 0 to 2^63-1. Code := 1 << Id

ExecutionExceptionId

Id type for execution exceptions. Id ranges from 0 to 63

RoutineResult

The required result type of a user-defined routine

RtoResult

Memory module result type

SpawnFn

The spawn function for creating new tasks

StateType