qor_rto::prelude
Struct Concurrency

pub struct Concurrency { /* private fields */ }

A Concurrency is an action that executes other actions in multiple branches concurrently. It completes when all concurrent actions have completed. One branch throwing an exception will cause the Concurrency to throw the exception as well.

Implementations

impl Concurrency

pub fn new() -> Box<Self>

Creates a new Concurrency action

pub fn with_branch(self: Box<Self>, action: Box<dyn Action>) -> Box<Self>

Adds a new action to the Concurrency at initialization

pub fn add_branch(&mut self, action: Box<dyn Action>) -> RtoResult<()>

Adds an action as branch to an existing Concurrency