qor_rto::prelude
Trait Listen

pub trait Listen: Debug + Send {
    // Required methods
    fn check(&self) -> bool;
    fn check_and_reset(&self) -> bool;
    fn wait(&self) -> RtoResult<()>;
    fn wait_timeout(&self, timeout: Duration) -> RtoResult<TimeoutStatus>;
}

Listen trait for event listeners.

Required Methods

fn check(&self) -> bool

Check if the subscribed event has received a notification since the last check.

Just checks on the event but leaves the status untouched.

fn check_and_reset(&self) -> bool

Check if the subscribed event has received a notification since the last check and reset it.

Reset the event notification state after checking.

fn wait(&self) -> RtoResult<()>

Wait for the event to trigger.

This locks the current thread until the event triggers. The event is reset when this function returns. If the event has already triggered (check would return true), the function returns immediately.

fn wait_timeout(&self, timeout: Duration) -> RtoResult<TimeoutStatus>

Wait for the event to trigger with a timeout.

This locks the current thread until the event triggers or the timeout is reached. The event is reset when this function returns. If the event has already triggered (check would return true), the function returns immediately.