qor_rto
Struct Computation

pub struct Computation { /* private fields */ }

A Computation is an action that executes multiple actions concurrently. It completes

with UpdateResult::Ready when all of the concurrent actions are UpdateState::Ready or

with UpdateResult::Complete one of them is UpdateState::Complete.

A branch throwing an exception will cause the Computation to throw the exception as well.

The Computation is a complex concurrent action. It can be used to model runtime races as well as repetitive calculations as required in generators or simulations.

Example: Execute a task and monitor it through a timeout

    // computation {
    //   await(my_task),
    //   throw(Timeout, 100ms)
    // }
    let Computation = Computation::new()
                        .with(Box::new(Await::new(async { /* my task */})))           // Await the asynchronous task
                        .with(Box::new(Throw::timeout(Duration::from_millis(100))));  // Race against a timeout

Explanation: The Computation will concurrently execute the asynchronous task and the Throw Timeout action with a timeout of 100ms. Both task and timeout will report UpdateState::Busy until the task completes or the timeout expires. If the task completes first the Computation will complete successfully. If the timeout expires first, the Computation will throw the timeout exception. Wrapping this action into a TryCatch action will allow to catch the timeout exception and handle it:

  // try {
  //    computation {
  //       await(my_task),
  //       throw(Timeout, 100ms)
  //    }
  // } catch(Timeout) {
  //    /* exception handling goes here */
  // }
  //
  let trycatch = TryCatch::for_timeout()
    .with_try(Box::new(Computation))
    .with_catch(Box::new(Await::new(async { /* handle timeout */ })));

   trycatch.init().unwrap();
   trycatch.start().unwrap();
   assert_eq!(trycatch.update(&Duration::from_millis(50), UpdateResult::Busy));
   assert_eq!(trycatch.update(&Duration::from_millis(50), UpdateResult::Complete));
   trycatch.finalize().unwrap();
   trycatch.terminate().unwrap();

Implementations

impl Computation

pub fn new() -> Box<Self>

Creates a new Concurrency action

pub fn with_branch(self: Box<Self>, action: Box<dyn Action>) -> Box<Self>

Adds a new action to the Computation at initialization

pub fn add_branch(&mut self, action: Box<dyn Action>) -> RtoResult<()>

Adds an action as branch to the Computation