qor_rto
Struct Engine

pub struct Engine { /* private fields */ }

An engine runs tasks concurrently on threads. It is the central executor of runtime orchestration.

The engine uses a threadpool to run tasks concurrently. Tasks are scheduled on the engine and run on the next available thread.

The engine can be obtained by either calling executor::default() or by utilizing an executor::EngineBuilder which gives control over the configuration of the engine.

The engine is compliant with Rust’s async executor model so the following is possible:

use qor_rto::{EngineBuilder, Task};

async fn my_async_function() -> () {
  // do something
}

fn main() {
  let executor = EngineBuilder::new().with_threads(4).build().unwrap();
  let task = Task::new(|| my_async_function());

  executor.start(); // start the engine
  let handle = executor.spawn(task); // schedule the task. the engine will put it on an available thread and run it.

  let result = handle.expect("valid handle").join(); // wait for the task to complete and get the result
  // do something with the result
}

Implementations

impl Engine

pub const MAX_THREADS: usize = 256usize

The maximum number of threads that can be used by the engine.

pub const MAX_TASKS: usize = 2_048usize

The maximum number of tasks that can be scheduled on the engine.

pub fn state(&self) -> ExecutionState

Get the current state of the engine.

pub fn is_ready(&self) -> bool

Check if the engine is ready to be started.

The engine is ready if it has been initialized or reset and has not reveived a stop signal.

pub fn is_running(&self) -> bool

Check if the engine is running and can execute tasks.

The engine is running if it has been started and has not received a stop signal.

pub fn is_terminating(&self) -> bool

Check if the engine is terminating.

The engine is terminating if it has received a stop signal and has not finished the shutdown process.

pub fn clock(&self) -> &dyn Clock

Get the clock that is used by the engine to measure time.

pub fn thread_capacity(&self) -> usize

Get the number of threads that can be used by the engine.

pub fn thread_utilization(&self) -> (usize, usize)

Get the utilization of the engine’s threads.

The tuple contains the number of working threads busy with a task and the number of running threads.

pub fn task_capacity(&self) -> usize

Get the number of tasks that can be scheduled on the engine.

pub fn task_utilization(&self) -> (usize, usize)

Get the utilization of the tasks spawned on the engine.

The tuple contains the number of tasks on working threads and the total number of tasks on the engine. This value is a snapshot and may change immediately after the call. It also may contain glitches as the call may happen when a task is in transit from scheduling queue into working thread.

pub fn idle(&self) -> bool

Checks if the engine is idle. The engine is idle if

The engine is running and no tasks are scheduled

The engine is not running

pub fn clear(&self) -> RtoResult<()>

Clear the task queue.

After this all scheduled tasks are removed from the engine. Running tasks are not affected by the clear operation.

pub fn init(&self) -> RtoResult<ExecutionState>

Initialize the engine.

This creates the thread pool and prepares the engine for task execution. In the engine implementation the init() is a separate function for two reasons:

It is in line with the exectution lifetime model of the ExecutionStateMachine.

It allows to defer the initialization from the start, significantly speeding up the engine start operation.

pub fn reset(&self) -> RtoResult<ExecutionState>

Reset the engine into post-initialization state.

This clears the task queue, cancels all running tasks and waits for the running tasks to finish. Reset cannot be called when the engine is uninitialized or terminated.

pub fn start(&self) -> RtoResult<ExecutionState>

Start the engine.

If the engine has not been initialized yet, this will automatically do so. This will start the threads and make the engine ready to run tasks. This function will return after the engine is started.

pub fn shutdown(&self) -> RtoResult<ExecutionState>

Send a stop signal to the engine. The engine will shut down running tasks and stop the threads.

pub fn spawn<T>(&self, task: Task<T>) -> RtoResult<TaskHandle<T>>
where T: Send + 'static,

Spawn a new task on the engine.

The task will be scheduled on the engine and run on the next available thread. If scheduling is successful this returns the TaskHandle to be used for joining the task and obtaining the result.