qor_rto
Trait Action

pub trait Action: Debug + Send {
    // Required methods
    fn state(&self) -> ExecutionState;
    fn init(&mut self) -> ExecutionState;
    fn start(&mut self, ctx: &mut ExecutionContext<'_>) -> ExecutionState;
    fn update(
        &mut self,
        delta: &Duration,
        ctx: &mut ExecutionContext<'_>,
    ) -> (Duration, UpdateResult);
    fn finalize(&mut self, ctx: &mut ExecutionContext<'_>) -> ExecutionState;
    fn terminate(&mut self) -> ExecutionState;
}

The Action trait describes the interface to an executable element we call action.

Required Methods

fn state(&self) -> ExecutionState

Returns the current state of the action

fn init(&mut self) -> ExecutionState

Initialize a new action with the given actionable. This is a valid operation when the action is Uninitialized or Terminated.

fn start(&mut self, ctx: &mut ExecutionContext<'_>) -> ExecutionState

Start the action. This is a valid operation after the action has been initialized Ready or stopped Completed.

fn update( &mut self, delta: &Duration, ctx: &mut ExecutionContext<'_>, ) -> (Duration, UpdateResult)

Update the action with the given time delta. This is a valid operation after the action has been started and is Running.

fn finalize(&mut self, ctx: &mut ExecutionContext<'_>) -> ExecutionState

Stop and finalize the action. This is a valid operation when the action is Running.

fn terminate(&mut self) -> ExecutionState

Terminate the action. This is a valid operation when the action is Ready, Completed or in Err state.