qor_rto
Trait Clock

pub trait Clock {
    // Required methods
    fn frequency(&self) -> u64;
    fn min(&self) -> Instant;
    fn now(&self) -> Instant;
    fn sleep_until(&self, time: Instant);

    // Provided methods
    fn period(&self) -> Duration { ... }
    fn age(&self) -> Duration { ... }
    fn sleep(&self, duration: Duration) { ... }
}

A clock provides the current timepoint and can hold the execution sleep for a given duration.

We use clocks to abstract access to time. This way we can decouple the execution time from the real time.

In the runtime orchestration, a clock is always steady and monotonic.

Required Methods

fn frequency(&self) -> u64

Get the frequency of the clock, in Hz.

fn min(&self) -> Instant

Get the minimum timepoint of the clock. This is the timepoint when the clock started.

Do not assume this to be zero, as the clock may have started at any absolute point in time.

fn now(&self) -> Instant

Get the current timepoint of the clock.

fn sleep_until(&self, time: Instant)

Sleep until the given timepoint.

It is implementation specific for the clock what happens here. The clock may suspend the thread, or just busy wait until the given timepoint. If the timepoint is in the past, the function returns immediately.

Provided Methods

fn period(&self) -> Duration

Get the period of the clock, in nanoseconds.

This is by default the inverse of the frequency.

fn age(&self) -> Duration

Get the age of the clock.

The age is the duration since the clock started.

fn sleep(&self, duration: Duration)

Sleep for the given duration.

It is implementation specific for the clock what happens here. The clock may suspend the thread, or just busy wait for the given duration. In simulation clocks, this may even do nothing and just advance the time.