qor_rto
Trait Variable

pub trait Variable<T: TypeTag + Debug + Send + Sync>: LValue<T> + RValue<T> { }

A Variable is a trait that represents a scalar variable storage with instant updates The type T must comply to the Qore Type System.